refactor: move elatic class to under program concern
